Agents-3/requirements.txt
flair==0.13.1
distro==1.9.0
groq==0.8.0
python-decouple==3.8
rouge==1.0.1
rouge-score==0.1.2
nltk==3.8.1
requests==2.32.3
BeautifulSoup4==4.12.3